Split Regex code into subfiles
Move Cache related code to Regex.Cache.cs

Move Match related code to Regex.Match.cs

Move Replace related code to Regex.Replace.cs

Move Split related code to Match.Split.cs

Move Timeout related code to Regex.Timeout.cs
