+++ src/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Regex.cs
@@ -24,38 +24,10 @@ namespace System.Text.RegularExpressions
     /// contains static methods that allow use of regular expressions without instantiating
     /// a Regex explicitly.
     /// </summary>
-    public class Regex : ISerializable
+    public partial class Regex : ISerializable
     {
         protected internal string pattern;                   // The string pattern provided
         protected internal RegexOptions roptions;            // the top-level options from the options string
-
-        // *********** Match timeout fields { ***********
-
-        // We need this because time is queried using Environment.TickCount for performance reasons
-        // (Environment.TickCount returns milliseconds as an int and cycles):
-        private static readonly TimeSpan MaximumMatchTimeout = TimeSpan.FromMilliseconds(int.MaxValue - 1);
-
-        // InfiniteMatchTimeout specifies that match timeout is switched OFF. It allows for faster code paths
-        // compared to simply having a very large timeout.
-        // We do not want to ask users to use System.Threading.Timeout.InfiniteTimeSpan as a parameter because:
-        //   (1) We do not want to imply any relation between having using a RegEx timeout and using multi-threading.
-        //   (2) We do not want to require users to take ref to a contract assembly for threading just to use RegEx.
-        //       There may in theory be a SKU that has RegEx, but no multithreading.
-        // We create a public Regex.InfiniteMatchTimeout constant, which for consistency uses the save underlying
-        // value as Timeout.InfiniteTimeSpan creating an implementation detail dependency only.
-        public static readonly TimeSpan InfiniteMatchTimeout = Timeout.InfiniteTimeSpan;
-
-        protected internal TimeSpan internalMatchTimeout;   // timeout for the execution of this regex
-
-        // During static initialisation of Regex we check
-        private const string DefaultMatchTimeout_ConfigKeyName = "REGEX_DEFAULT_MATCH_TIMEOUT";
-
-        // DefaultMatchTimeout specifies the match timeout to use if no other timeout was specified
-        // by one means or another. Typically, it is set to InfiniteMatchTimeout.
-        internal static readonly TimeSpan DefaultMatchTimeout = InitDefaultMatchTimeout();
-
-        // *********** } match timeout fields ***********
-
         protected internal RegexRunnerFactory factory;
 
         protected internal Hashtable caps;          // if captures are sparse, this is the hashtable capnum->index
@@ -110,14 +82,11 @@ namespace System.Text.RegularExpressions
         internal RegexCode _code;                           // if interpreted, this is the code for RegexInterpreter
         internal bool _refsInitialized = false;
 
-        internal static LinkedList<CachedCodeEntry> s_livecode = new LinkedList<CachedCodeEntry>();// the cache of code and factories that are currently loaded
-        internal static int s_cacheSize = 15;
-
         internal const int MaxOptionShift = 10;
 
         protected Regex()
         {
-            internalMatchTimeout = DefaultMatchTimeout;
+            internalMatchTimeout = s_defaultMatchTimeout;
         }
 
         /// <summary>
@@ -125,7 +94,7 @@ namespace System.Text.RegularExpressions
         /// expression.
         /// </summary>
         public Regex(string pattern)
-            : this(pattern, RegexOptions.None, DefaultMatchTimeout, false)
+            : this(pattern, RegexOptions.None, s_defaultMatchTimeout, false)
         {
         }
 
@@ -134,7 +103,7 @@ namespace System.Text.RegularExpressions
         /// specified regular expression with options that modify the pattern.
         /// </summary>
         public Regex(string pattern, RegexOptions options)
-            : this(pattern, options, DefaultMatchTimeout, false)
+            : this(pattern, options, s_defaultMatchTimeout, false)
         {
         }
 
@@ -245,65 +214,6 @@ namespace System.Text.RegularExpressions
 #endif
         }
 
-        // Note: "&lt;" is the XML entity for smaller ("<").
-        /// <summary>
-        /// Validates that the specified match timeout value is valid.
-        /// The valid range is <code>TimeSpan.Zero &lt; matchTimeout &lt;= Regex.MaximumMatchTimeout</code>.
-        /// </summary>
-        /// <param name="matchTimeout">The timeout value to validate.</param>
-        /// <exception cref="ArgumentOutOfRangeException">If the specified timeout is not within a valid range.
-        /// </exception>
-        protected internal static void ValidateMatchTimeout(TimeSpan matchTimeout)
-        {
-            if (InfiniteMatchTimeout == matchTimeout)
-                return;
-
-            // Change this to make sure timeout is not longer then Environment.Ticks cycle length:
-            if (TimeSpan.Zero < matchTimeout && matchTimeout <= MaximumMatchTimeout)
-                return;
-
-            throw new ArgumentOutOfRangeException(nameof(matchTimeout));
-        }
-
-        /// <summary>
-        /// Specifies the default RegEx matching timeout value (i.e. the timeout that will be used if no
-        /// explicit timeout is specified).
-        /// The default is queried from the current <code>AppDomain</code>.
-        /// If the AddDomain's data value for that key is not a <code>TimeSpan</code> value or if it is outside the
-        /// valid range, an exception is thrown.
-        /// If the AddDomain's data value for that key is <code>null</code>, a fallback value is returned.
-        /// </summary>
-        /// <returns>The default RegEx matching timeout for this AppDomain</returns>
-        private static TimeSpan InitDefaultMatchTimeout()
-        {
-            // Query AppDomain
-            AppDomain ad = AppDomain.CurrentDomain;
-            object defaultMatchTimeoutObj = ad.GetData(DefaultMatchTimeout_ConfigKeyName);
-
-            // If no default is specified, use fallback
-            if (defaultMatchTimeoutObj == null)
-            {
-                return InfiniteMatchTimeout;
-            }
-
-            if (defaultMatchTimeoutObj is TimeSpan defaultMatchTimeOut)
-            {
-                // If default timeout is outside the valid range, throw. It will result in a TypeInitializationException:
-                try
-                {
-                    ValidateMatchTimeout(defaultMatchTimeOut);
-                }
-                catch (ArgumentOutOfRangeException)
-                {
-                    throw new ArgumentOutOfRangeException(SR.Format(SR.IllegalDefaultRegexMatchTimeoutInAppDomain, DefaultMatchTimeout_ConfigKeyName, defaultMatchTimeOut));
-                }
-
-                return defaultMatchTimeOut;
-            }
-
-            throw new InvalidCastException(SR.Format(SR.IllegalDefaultRegexMatchTimeoutInAppDomain, DefaultMatchTimeout_ConfigKeyName, defaultMatchTimeoutObj));
-        }
-
 #if FEATURE_COMPILED
         /// <summary>
         /// This method is here for perf reasons: if the call to RegexCompiler is NOT in the 
@@ -344,30 +254,6 @@ namespace System.Text.RegularExpressions
                 throw new ArgumentNullException(nameof(str));
 
             return RegexParser.Unescape(str);
-        }
-
-        [SuppressMessage("Microsoft.Concurrency", "CA8001", Justification = "Reviewed for thread-safety")]
-        public static int CacheSize
-        {
-            get
-            {
-                return s_cacheSize;
-            }
-            set
-            {
-                if (value < 0)
-                    throw new ArgumentOutOfRangeException(nameof(value));
-
-                s_cacheSize = value;
-                if (s_livecode.Count > s_cacheSize)
-                {
-                    lock (s_livecode)
-                    {
-                        while (s_livecode.Count > s_cacheSize)
-                            s_livecode.RemoveLast();
-                    }
-                }
-            }
         }
 
         /// <summary>
@@ -558,392 +444,6 @@ namespace System.Text.RegularExpressions
             return -1;
         }
 
-        /*
-         * Static version of simple IsMatch call
-         */
-        /// <summary>
-        /// Searches the input string for one or more occurrences of the text supplied in the given pattern.
-        /// </summary>
-        public static bool IsMatch(string input, string pattern)
-        {
-            return IsMatch(input, pattern, RegexOptions.None, DefaultMatchTimeout);
-        }
-
-        /*
-         * Static version of simple IsMatch call
-         */
-        /// <summary>
-        /// Searches the input string for one or more occurrences of the text
-        /// supplied in the pattern parameter with matching options supplied in the options
-        /// parameter.
-        /// </summary>
-        public static bool IsMatch(string input, string pattern, RegexOptions options)
-        {
-            return IsMatch(input, pattern, options, DefaultMatchTimeout);
-        }
-
-        public static bool IsMatch(string input, string pattern, RegexOptions options, TimeSpan matchTimeout)
-        {
-            return new Regex(pattern, options, matchTimeout, true).IsMatch(input);
-        }
-
-        /*
-         * Returns true if the regex finds a match within the specified string
-         */
-        /// <summary>
-        /// Searches the input string for one or more matches using the previous pattern,
-        /// options, and starting position.
-        /// </summary>
-        public bool IsMatch(string input)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return IsMatch(input, UseOptionR() ? input.Length : 0);
-        }
-
-        /*
-         * Returns true if the regex finds a match after the specified position
-         * (proceeding leftward if the regex is leftward and rightward otherwise)
-         */
-        /// <summary>
-        /// Searches the input string for one or more matches using the previous pattern and options,
-        /// with a new starting position.
-        /// </summary>
-        public bool IsMatch(string input, int startat)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return (null == Run(true, -1, input, 0, input.Length, startat));
-        }
-
-        /*
-         * Static version of simple Match call
-         */
-        /// <summary>
-        /// Searches the input string for one or more occurrences of the text
-        /// supplied in the pattern parameter.
-        /// </summary>
-        public static Match Match(string input, string pattern)
-        {
-            return Match(input, pattern, RegexOptions.None, DefaultMatchTimeout);
-        }
-
-        /*
-         * Static version of simple Match call
-         */
-        /// <summary>
-        /// Searches the input string for one or more occurrences of the text
-        /// supplied in the pattern parameter. Matching is modified with an option
-        /// string.
-        /// </summary>
-        public static Match Match(string input, string pattern, RegexOptions options)
-        {
-            return Match(input, pattern, options, DefaultMatchTimeout);
-        }
-
-
-        public static Match Match(string input, string pattern, RegexOptions options, TimeSpan matchTimeout)
-        {
-            return new Regex(pattern, options, matchTimeout, true).Match(input);
-        }
-
-        /*
-         * Finds the first match for the regular expression starting at the beginning
-         * of the string (or at the end of the string if the regex is leftward)
-         */
-        /// <summary>
-        /// Matches a regular expression with a string and returns
-        /// the precise result as a RegexMatch object.
-        /// </summary>
-        public Match Match(string input)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return Match(input, UseOptionR() ? input.Length : 0);
-        }
-
-        /*
-         * Finds the first match, starting at the specified position
-         */
-        /// <summary>
-        /// Matches a regular expression with a string and returns
-        /// the precise result as a RegexMatch object.
-        /// </summary>
-        public Match Match(string input, int startat)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return Run(false, -1, input, 0, input.Length, startat);
-        }
-
-        /*
-         * Finds the first match, restricting the search to the specified interval of
-         * the char array.
-         */
-        /// <summary>
-        /// Matches a regular expression with a string and returns the precise result as a
-        /// RegexMatch object.
-        /// </summary>
-        public Match Match(string input, int beginning, int length)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return Run(false, -1, input, beginning, length, UseOptionR() ? beginning + length : beginning);
-        }
-
-        /*
-         * Static version of simple Matches call
-         */
-        /// <summary>
-        /// Returns all the successful matches as if Match were called iteratively numerous times.
-        /// </summary>
-        public static MatchCollection Matches(string input, string pattern)
-        {
-            return Matches(input, pattern, RegexOptions.None, DefaultMatchTimeout);
-        }
-
-        /*
-         * Static version of simple Matches call
-         */
-        /// <summary>
-        /// Returns all the successful matches as if Match were called iteratively numerous times.
-        /// </summary>
-        public static MatchCollection Matches(string input, string pattern, RegexOptions options)
-        {
-            return Matches(input, pattern, options, DefaultMatchTimeout);
-        }
-
-        public static MatchCollection Matches(string input, string pattern, RegexOptions options, TimeSpan matchTimeout)
-        {
-            return new Regex(pattern, options, matchTimeout, true).Matches(input);
-        }
-
-        /*
-         * Finds the first match for the regular expression starting at the beginning
-         * of the string Enumerator(or at the end of the string if the regex is leftward)
-         */
-        /// <summary>
-        /// Returns all the successful matches as if Match was called iteratively numerous times.
-        /// </summary>
-        public MatchCollection Matches(string input)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return Matches(input, UseOptionR() ? input.Length : 0);
-        }
-
-        /*
-         * Finds the first match, starting at the specified position
-         */
-        /// <summary>
-        /// Returns all the successful matches as if Match was called iteratively numerous times.
-        /// </summary>
-        public MatchCollection Matches(string input, int startat)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return new MatchCollection(this, input, 0, input.Length, startat);
-        }
-
-        /// <summary>
-        /// Replaces all occurrences of the pattern with the <paramref name="replacement"/> pattern, starting at
-        /// the first character in the input string.
-        /// </summary>
-        public static string Replace(string input, string pattern, string replacement)
-        {
-            return Replace(input, pattern, replacement, RegexOptions.None, DefaultMatchTimeout);
-        }
-
-        /// <summary>
-        /// Replaces all occurrences of
-        /// the <paramref name="pattern "/>with the <paramref name="replacement "/>
-        /// pattern, starting at the first character in the input string.
-        /// </summary>
-        public static string Replace(string input, string pattern, string replacement, RegexOptions options)
-        {
-            return Replace(input, pattern, replacement, options, DefaultMatchTimeout);
-        }
-
-        public static string Replace(string input, string pattern, string replacement, RegexOptions options, TimeSpan matchTimeout)
-        {
-            return new Regex(pattern, options, matchTimeout, true).Replace(input, replacement);
-        }
-
-        /// <summary>
-        /// Replaces all occurrences of the previously defined pattern with the
-        /// <paramref name="replacement"/> pattern, starting at the first character in the
-        /// input string.
-        /// </summary>
-        public string Replace(string input, string replacement)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return Replace(input, replacement, -1, UseOptionR() ? input.Length : 0);
-        }
-
-        /// <summary>
-        /// Replaces all occurrences of the previously defined pattern with the
-        /// <paramref name="replacement"/> pattern, starting at the first character in the
-        /// input string.
-        /// </summary>
-        public string Replace(string input, string replacement, int count)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return Replace(input, replacement, count, UseOptionR() ? input.Length : 0);
-        }
-
-        /// <summary>
-        /// Replaces all occurrences of the previously defined pattern with the
-        /// <paramref name="replacement"/> pattern, starting at the character position
-        /// <paramref name="startat"/>.
-        /// </summary>
-        public string Replace(string input, string replacement, int count, int startat)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            if (replacement == null)
-                throw new ArgumentNullException(nameof(replacement));
-
-            // a little code to grab a cached parsed replacement object
-            RegexReplacement repl = (RegexReplacement)_replref.Get();
-
-            if (repl == null || !repl.Pattern.Equals(replacement))
-            {
-                repl = RegexParser.ParseReplacement(replacement, caps, capsize, capnames, roptions);
-                _replref.Cache(repl);
-            }
-
-            return repl.Replace(this, input, count, startat);
-        }
-
-        /// <summary>
-        /// Replaces all occurrences of the <paramref name="pattern"/> with the recent
-        /// replacement pattern.
-        /// </summary>
-        public static string Replace(string input, string pattern, MatchEvaluator evaluator)
-        {
-            return Replace(input, pattern, evaluator, RegexOptions.None, DefaultMatchTimeout);
-        }
-
-        /// <summary>
-        /// Replaces all occurrences of the <paramref name="pattern"/> with the recent
-        /// replacement pattern, starting at the first character.
-        /// </summary>
-        public static string Replace(string input, string pattern, MatchEvaluator evaluator, RegexOptions options)
-        {
-            return Replace(input, pattern, evaluator, options, DefaultMatchTimeout);
-        }
-
-        public static string Replace(string input, string pattern, MatchEvaluator evaluator, RegexOptions options, TimeSpan matchTimeout)
-        {
-            return new Regex(pattern, options, matchTimeout, true).Replace(input, evaluator);
-        }
-
-        /// <summary>
-        /// Replaces all occurrences of the previously defined pattern with the recent
-        /// replacement pattern, starting at the first character position.
-        /// </summary>
-        public string Replace(string input, MatchEvaluator evaluator)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return Replace(input, evaluator, -1, UseOptionR() ? input.Length : 0);
-        }
-
-        /// <summary>
-        /// Replaces all occurrences of the previously defined pattern with the recent
-        /// replacement pattern, starting at the first character position.
-        /// </summary>
-        public string Replace(string input, MatchEvaluator evaluator, int count)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return Replace(input, evaluator, count, UseOptionR() ? input.Length : 0);
-        }
-
-        /// <summary>
-        /// Replaces all occurrences of the previously defined pattern with the recent
-        /// replacement pattern, starting at the character position
-        /// <paramref name="startat"/>.
-        /// </summary>
-        public string Replace(string input, MatchEvaluator evaluator, int count, int startat)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return RegexReplacement.Replace(evaluator, this, input, count, startat);
-        }
-
-        /// <summary>
-        /// Splits the <paramref name="input "/>string at the position defined
-        /// by <paramref name="pattern"/>.
-        /// </summary>
-        public static string[] Split(string input, string pattern)
-        {
-            return Split(input, pattern, RegexOptions.None, DefaultMatchTimeout);
-        }
-
-        /// <summary>
-        /// Splits the <paramref name="input "/>string at the position defined by <paramref name="pattern"/>.
-        /// </summary>
-        public static string[] Split(string input, string pattern, RegexOptions options)
-        {
-            return Split(input, pattern, options, DefaultMatchTimeout);
-        }
-
-        public static string[] Split(string input, string pattern, RegexOptions options, TimeSpan matchTimeout)
-        {
-            return new Regex(pattern, options, matchTimeout, true).Split(input);
-        }
-
-        /// <summary>
-        /// Splits the <paramref name="input"/> string at the position defined by a
-        /// previous pattern.
-        /// </summary>
-        public string[] Split(string input)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return Split(input, 0, UseOptionR() ? input.Length : 0);
-        }
-
-        /// <summary>
-        /// Splits the <paramref name="input"/> string at the position defined by a
-        /// previous pattern.
-        /// </summary>
-        public string[] Split(string input, int count)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return RegexReplacement.Split(this, input, count, UseOptionR() ? input.Length : 0);
-        }
-
-        /// <summary>
-        /// Splits the <paramref name="input"/> string at the position defined by a previous pattern.
-        /// </summary>
-        public string[] Split(string input, int count, int startat)
-        {
-            if (input == null)
-                throw new ArgumentNullException(nameof(input));
-
-            return RegexReplacement.Split(this, input, count, startat);
-        }
-
 #if FEATURE_COMPILED
         [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "assemblyname", Justification = "Microsoft: already shipped since v1 - can't fix without causing a breaking change")]
         public static void CompileToAssembly(RegexCompilationInfo[] regexinfos, AssemblyName assemblyname)
@@ -1023,61 +523,6 @@ namespace System.Text.RegularExpressions
             return match;
         }
 
-        /*
-         * Find code cache based on options+pattern
-         */
-        private static CachedCodeEntry LookupCachedAndUpdate(CachedCodeEntryKey key)
-        {
-            lock (s_livecode)
-            {
-                for (LinkedListNode<CachedCodeEntry> current = s_livecode.First; current != null; current = current.Next)
-                {
-                    if (current.Value._key == key)
-                    {
-                        // If we find an entry in the cache, move it to the head at the same time.
-                        s_livecode.Remove(current);
-                        s_livecode.AddFirst(current);
-                        return current.Value;
-                    }
-                }
-            }
-
-            return null;
-        }
-
-        /*
-         * Add current code to the cache
-         */
-        private CachedCodeEntry CacheCode(CachedCodeEntryKey key)
-        {
-            CachedCodeEntry newcached = null;
-
-            lock (s_livecode)
-            {
-                // first look for it in the cache and move it to the head
-                for (LinkedListNode<CachedCodeEntry> current = s_livecode.First; current != null; current = current.Next)
-                {
-                    if (current.Value._key == key)
-                    {
-                        s_livecode.Remove(current);
-                        s_livecode.AddFirst(current);
-                        return current.Value;
-                    }
-                }
-
-                // it wasn't in the cache, so we'll add a new one.  Shortcut out for the case where cacheSize is zero.
-                if (s_cacheSize != 0)
-                {
-                    newcached = new CachedCodeEntry(key, capnames, capslist, _code, caps, capsize, _runnerref, _replref);
-                    s_livecode.AddFirst(newcached);
-                    if (s_livecode.Count > s_cacheSize)
-                        s_livecode.RemoveLast();
-                }
-            }
-
-            return newcached;
-        }
-
         protected bool UseOptionC()
         {
             return (roptions & RegexOptions.Compiled) != 0;
@@ -1106,94 +551,6 @@ namespace System.Text.RegularExpressions
             {
                 return (roptions & RegexOptions.Debug) != 0;
             }
-        }
-#endif
-    }
-
-
-    /*
-     * Callback class
-     */
-    public delegate string MatchEvaluator(Match match);
-
-    /*
-     * Used as a key for CacheCodeEntry
-     */
-    internal readonly struct CachedCodeEntryKey : IEquatable<CachedCodeEntryKey>
-    {
-        private readonly RegexOptions _options;
-        private readonly string _cultureKey;
-        private readonly string _pattern;
-
-        internal CachedCodeEntryKey(RegexOptions options, string cultureKey, string pattern)
-        {
-            _options = options;
-            _cultureKey = cultureKey;
-            _pattern = pattern;
-        }
-
-        public override bool Equals(object obj)
-        {
-            return obj is CachedCodeEntryKey && Equals((CachedCodeEntryKey)obj);
-        }
-
-        public bool Equals(CachedCodeEntryKey other)
-        {
-            return this == other;
-        }
-
-        public static bool operator ==(CachedCodeEntryKey left, CachedCodeEntryKey right)
-        {
-            return left._options == right._options && left._cultureKey == right._cultureKey && left._pattern == right._pattern;
-        }
-
-        public static bool operator !=(CachedCodeEntryKey left, CachedCodeEntryKey right)
-        {
-            return !(left == right);
-        }
-
-        public override int GetHashCode()
-        {
-            return ((int)_options) ^ _cultureKey.GetHashCode() ^ _pattern.GetHashCode();
-        }
-    }
-
-    /*
-     * Used to cache byte codes
-     */
-    internal sealed class CachedCodeEntry
-    {
-        internal CachedCodeEntryKey _key;
-        internal RegexCode _code;
-        internal Hashtable _caps;
-        internal Hashtable _capnames;
-        internal string[] _capslist;
-#if FEATURE_COMPILED
-        internal RegexRunnerFactory _factory;
-#endif
-        internal int _capsize;
-        internal ExclusiveReference _runnerref;
-        internal SharedReference _replref;
-
-        internal CachedCodeEntry(CachedCodeEntryKey key, Hashtable capnames, string[] capslist, RegexCode code, Hashtable caps, int capsize, ExclusiveReference runner, SharedReference repl)
-        {
-            _key = key;
-            _capnames = capnames;
-            _capslist = capslist;
-
-            _code = code;
-            _caps = caps;
-            _capsize = capsize;
-
-            _runnerref = runner;
-            _replref = repl;
-        }
-
-#if FEATURE_COMPILED
-        internal void AddCompiled(RegexRunnerFactory factory)
-        {
-            _factory = factory;
-            _code = null;
         }
 #endif
     }
