+++ src/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexCompilationInfo.cs
@@ -12,7 +12,7 @@ namespace System.Text.RegularExpressions
         private TimeSpan _matchTimeout;
 
         public RegexCompilationInfo(string pattern, RegexOptions options, string name, string fullnamespace, bool ispublic)
-            : this(pattern, options, name, fullnamespace, ispublic, Regex.DefaultMatchTimeout)
+            : this(pattern, options, name, fullnamespace, ispublic, Regex.s_defaultMatchTimeout)
         {
         }
